requirements.txt
requests==2.32.5
python-dotenv==1.2.2
customtkinter==5.2.2